# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_hierarchy_imports_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "A .csv or .xlsx of the structure.",
            "title": "File",
            "type": "string"
          },
          "sheet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Name"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_hierarchy_imports_post",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, safe to branch on.",
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "default": "",
            "title": "Correlation Id",
            "type": "string"
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "description": "Safe to display to a person.",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "ImportApply": {
        "additionalProperties": false,
        "description": "The version the person was looking at when they decided.\n\nNot decoration: between reading the preview and pressing apply, somebody else may have\nre-mapped the columns. Applying a tree the person never saw is the failure this prevents.",
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "ImportApply",
        "type": "object"
      },
...
      "ImportSummary": {
        "description": "Where an import has got to, and what it made of the file.",
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "column_mapping": {
            "additionalProperties": true,
            "title": "Column Mapping",
            "type": "object"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ignored_columns": {
            "items": {
              "type": "string"
            },
            "title": "Ignored Columns",
            "type": "array"
          },
          "proposal": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "sheet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Name"
          },
          "source_columns": {
            "items": {
              "type": "string"
            },
            "title": "Source Columns",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "warning_count": {
            "title": "Warning Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "sheet_name",
          "source_columns",
          "column_mapping",
          "ignored_columns",
          "proposal",
          "row_count",
          "error_count",
          "warning_count",
          "applied_at",
          "version"
        ],
        "title": "ImportSummary",
        "type": "object"
      },
...
      "TreeRead": {
        "description": "The whole tree, flat.\n\nFlat rather than nested, deliberately. The client builds the nesting from `parent_id`, which\nmeans one shape serves the tree view, a search result and a partially expanded view without\nthe server guessing which is wanted. A nested response would also have to pick a depth limit,\nand any limit is wrong for somebody.",
        "properties": {
          "as_at": {
            "format": "date",
            "title": "As At",
            "type": "string"
          },
          "is_empty": {
            "title": "Is Empty",
            "type": "boolean"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/OrgUnitRead"
            },
            "title": "Units",
            "type": "array"
          }
        },
        "required": [
          "as_at",
          "units",
          "is_empty"
        ],
        "title": "TreeRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Governed human and AI work operating system",
    "title": "UBOSS AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hierarchy": {
      "get": {
        "description": "Units and positions, flat, as at one date.\n\nFlat rather than nested: the client builds the nesting from `parent_id`, which means one\nresponse shape serves the tree, a search result and a partially expanded view. A nested\nresponse would also have to pick a depth limit, and every limit is wrong for somebody.\n\n`is_empty` distinguishes \"this organisation has no tree yet\" from \"the request failed\" \u2014 two\nstates that look identical on screen and mean opposite things.",
        "operationId": "read_tree_api_v1_hierarchy_get",
        "parameters": [
          {
            "description": "Show the structure as it stands on this date. Defaults to today.",
            "in": "query",
            "name": "as_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show the structure as it stands on this date. Defaults to today.",
              "title": "As At"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "The company tree",
        "tags": [
          "hierarchy"
        ]
      }
    },
...
    "/api/v1/hierarchy/imports": {
      "post": {
        "description": "PLAN \u00a75 steps 1 and 2 \u2014 quarantine, then parse deterministically.\n\nNothing is created in the live tree. The file is stored `pending` and stays that way: an\nimport source is never served back to a browser, so it never needs to leave quarantine.\n\nThe response says which columns were understood and which were not. The ones that were not\nare what step 3 may ask a model about.",
        "operationId": "upload_api_v1_hierarchy_imports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_hierarchy_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "Upload a structure file",
        "tags": [
          "hierarchy"
        ]
      }
    },
    "/api/v1/hierarchy/imports/{import_id}": {
      "get": {
        "description": "PLAN \u00a75 step 5 \u2014 row errors, warnings, and the proposed tree.\n\nThe tree is derived from the staged rows on every read rather than stored. Two copies of one\nfact drift, and the copy on screen is the one somebody would act on.",
        "operationId": "preview_api_v1_hierarchy_imports__import_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "The staged rows and the tree they would build",
        "tags": [
          "hierarchy"
        ]
      }
    },
    "/api/v1/hierarchy/imports/{import_id}/apply": {
      "post": {
        "description": "PLAN \u00a75 step 7 \u2014 *\"Backend applies atomically and records source/mapping/audit.\"*\n\nOne transaction: either the whole staged tree exists afterwards or none of it does. A\nhalf-applied org chart is worse than a failed one, because nobody can tell which half is real.\n\nRefused while any row has an error. An import applied with known bad rows puts known bad data\ninto the thing every permission scope reads from.",
        "operationId": "apply_api_v1_hierarchy_imports__import_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportApply"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "Apply the import",
        "tags": [
          "hierarchy"
        ]
      }
    },
...
    "/api/v1/hierarchy/imports/{import_id}/propose": {
      "post": {
        "description": "PLAN \u00a75 step 3 \u2014 *\"Claude proposes only ambiguous column mappings.\"*\n\nOnly the headings the deterministic pass could not place, and only headings: the model never\nsees the tree, is never asked what to create, and its answer is a suggestion until a person\naccepts it.\n\nReturns 200 whether or not a model was reachable. `proposal.consulted` says which happened,\nso the screen can state it rather than leaving a person to infer it from an empty list.",
        "operationId": "propose_api_v1_hierarchy_imports__import_id__propose_post",
        "parameters": [
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Import Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not signed in, or the session ended."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused. The reason is in the audit trail."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such record, for this caller."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The record moved, or the key was reused."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Some field was not accepted."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too many attempts. See Retry-After."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A fault on our side. Nothing was changed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A dependency did not answer. Retryable."
          }
        },
        "summary": "Ask a model about the unmatched columns",
        "tags": [
          "hierarchy"
        ]
      }
    },
...
  }
}